refactor: companyLogoMap.js 디렉토리 변경
