+++ src/constants/companyLogoMap.js
